requirements.txt
aiolimiter==1.1.1
anyio==4.9.0
APScheduler==3.10.4
cachetools==5.3.3
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-dotenv==1.1.0
python-telegram-bot==21.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
typing_extensions==4.13.1
tzlocal==5.3.1
